requirements.txt
fastapi
uvicorn
joblib
tensorflow
pillow
numpy
requests
